requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
future==0.17.1
google==2.0.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.6
grpcio==1.17.1
idna==2.8
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.7
termcolor==1.1.0
urllib3==1.24.1
